refactor: changed no match errors to emit on a subcommand listener instead of the root MessageCommandError / ChatInputCommandError
BREAKING CHANGE: Previously if no subcommand matched the error was emitted to `MessageCommandError` or `ChatInputCommandError`.
This was inconsistent with errors during runtime being emitted to `MessageSubcommandError` and `ChatInputSubcommandError` respectively.
Therefore these events are now emitted to the new events `PluginMessageSubcommandNoMatch` and `PluginChatInputSubcommandNoMatch` respectively.
There are new built-in listeners for these events that will log the infractions to the console at error level. You can override these listeners to provide your own functionality.