refactor: changed no match errors to emit on a subcommand listener instead of the root MessageCommandError / ChatInputCommandError (#601)
BREAKING CHANGE: If no subcommand matched the error was emitted to `MessageCommandError` or `ChatInputCommandError`.
This was inconsistent with errors during runtime emitted to `MessageSubcommandError` and `ChatInputSubcommandError` respectively.
Therefore these events are now emitted to the new events `PluginMessageSubcommandNoMatch` and `PluginChatInputSubcommandNoMatch` respectively.
New built-in listeners for these events will log the infractions to the console at the error level. You can override these listeners to provide your functionality.